Reorganize code (more modular), remove flaky test
- Move main logic from a single file to a dedicated llm_tools_mcp directory.
- Split code into mcp_client, mcp_config, and register_tools modules.
- Update entry point and test imports to reflect the new structure.
- Remove flaky postman-echo integration test.
