+++ tests/test_config_validation.py
@@ -6,7 +6,7 @@ import pytest
 import uvicorn
 from starlette.applications import Starlette
 
-from llm_tools_mcp import McpConfig
+from llm_tools_mcp.mcp_config import McpConfig
 
 from mcp.server.fastmcp import FastMCP
