+++ tests/test_integration.py
@@ -1,42 +1,7 @@
 import json
 import pytest
-from llm_tools_mcp import McpConfig, McpClient
-
-
-@pytest.mark.asyncio
-@pytest.mark.online
-async def test_streamable_http_postman_echo_mcp():
-    """Test streamable HTTP connection to postman-echo-mcp service and list its echoes."""
-    mcp_config_content = json.dumps(
-        {
-            "mcpServers": {
-                "postman-echo": {
-                    "type": "http",
-                    "url": "https://postman-echo-mcp.fly.dev/",
-                }
-            }
-        }
-    )
-
-    mcp_config_obj = McpConfig.for_json_content(mcp_config_content)
-    mcp_client = McpClient(mcp_config_obj)
-
-    tools = await mcp_client.get_all_tools()
-
-    assert "postman-echo" in tools, "Should have postman-echo server"
-
-    tools = tools.get("postman-echo", [])
-    tool_names = [tool.name for tool in tools]
-
-    assert "echo" in tool_names, f"Should have 'echo' tool. Found tools: {tool_names}"
-
-    result = await mcp_client.call_tool(
-        "postman-echo", "echo", message="Hello from test!"
-    )
-    assert result is not None, "Tool call should return a result"
-    assert "Hello from test!" in str(result), (
-        "Echo result should contain the original message"
-    )
+from llm_tools_mcp.mcp_client import McpClient
+from llm_tools_mcp.mcp_config import McpConfig
 
 
 @pytest.mark.asyncio
